upd finishing project
