+++ practice/src/components/button/Button.css
@@ -8,4 +8,5 @@
     height: 30px;
     font-size: 20px;
     display: block;
+    cursor: pointer;
 }
