+++ practice/src/components/footer/footer.css
@@ -1,16 +1,17 @@
+.footer {
+    margin-top: auto;
+}
+
 .footerContent {
     background-color: grey;
     color: white;
     width: 100%;
-    height: 100%;
-    display:block;
-    padding: 5px;
 }
 
 .footerLogo {
     max-width: 100px;
     max-height: 100px;
-    margin-left: 5px;
+    margin-left: 10px;
 }
 
 a {
@@ -18,7 +19,7 @@ a {
 }
 
 .footerLink {
-    margin: 5px;
+    margin-left: 10px;
 }
 
 .footerLink:hover {
@@ -27,5 +28,5 @@ a {
 }
 
 .footerData {
-    margin: 5px;
+    margin-left: 10px;
 }
