+++ practice/src/components/navigation/Navigation.css
@@ -1,13 +1,39 @@
 .navigationClick {
     cursor: pointer;
+    max-width: 1500px;
+    margin:auto;
+    background-color: #ffffff;
+    border-radius: 15px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    padding: 20px;
+    margin-bottom: 15px;
+}
+
+.navigationCharacter {
+    display: flex;
+    align-items: center;
+    margin-bottom: 20px;
+    border-bottom: 1px solid #ddd;
+    padding-bottom: 20px;    
 }
 
 .navigationImage {
-    width: 400px;
-    height: 300px;
+    max-width: 150px;
+    max-height: 150px;
+    border-radius: 15px;
+    margin-right: 20px;
 }
 
 .navigationTitle {
-    color: red;
-    font-weight: 400;
+    width: 20%;
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+    font-size: 25px;
+    color: #e74c3c;
+}
+
+.navigationDescription {
+    text-align: left;
+    width: 70%;
+    margin-left: 25px;
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
 }
