+++ practice/src/pages/Characters/Characters.css
@@ -1,0 +1,7 @@
+.charactersCounter {
+     color: rgb(128, 128, 128);
+}
+
+.charactersTitle {
+     margin-left: 10px;
+}
