+++ practice/src/pages/Comics/Comics.css
@@ -1,0 +1,7 @@
+.comicsCounter {
+    color: rgb(128, 128, 128);
+}
+
+.comicsTitle {
+    padding-left: 10px;
+}
